Moved stuff into a modals/ folder, and other stuff into an applications/ folder
