+++ script/mkcomp
@@ -53,7 +53,7 @@ import { $, Expression, Executor, Dataset } from 'plywood';
 import { Stage, Clicker, Essence, DataCube, Filter, Dimension, Measure } from '../../../common/models/index';
 // import { ... } from '../../config/constants';
 // import { Fn } from '../../../common/utils/general/general';
-import { SvgIcon } from '../svg-icon/svg-icon';
+import { SvgIcon } from '../index';
 
 // I am: import { ${camelName} } from '../${name}/${name}';
