+++ src/client/applications/no-data-cubes-application/no-data-cubes-application.scss
