+++ src/client/applications/pivot-application/pivot-application.scss
