+++ src/client/modals/about-modal/about-modal.scss
