+++ src/client/modals/add-collection-modal/add-collection-modal.scss
