+++ src/client/modals/add-collection-tile-modal/add-collection-tile-modal.scss
