+++ src/client/modals/index.ts
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2015-2016 Imply Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export * from './about-modal/about-modal';
+export * from './add-collection-modal/add-collection-modal';
+export * from './add-collection-tile-modal/add-collection-tile-modal';
+export * from './dimension-modal/dimension-modal';
+export * from './measure-modal/measure-modal';
+export * from './raw-data-modal/raw-data-modal';
+export * from './remove-modal/remove-modal';
