+++ src/client/modals/measure-modal/measure-modal.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-@import '../../../imports';
+@import '../../imports';
 
 .measure-modal {
   form {
