+++ src/client/utils/dom-highlighter/dom-highlighter.ts
